requirements.txt
Flask
python-magic
libmagic